+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/Config.java
@@ -48,12 +48,9 @@ public class Config {
     public boolean verbose = false;
     public boolean copyOriginalFiles = false;
     public boolean updateFiles = false;
-    public boolean isFramework = false;
-    public boolean resourcesAreCompressed = false;
     public boolean useAapt2 = false;
     public boolean noCrunch = false;
     public int forceApi = 0;
-    public Collection<String> doNotCompress;
 
     // Decode options
     public short decodeSources = DECODE_SOURCES_SMALI;
