+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/apk/PackageInfo.java
@@ -14,11 +14,9 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package brut.androlib.meta;
+package brut.androlib.apk;
 
-import java.util.List;
-
-public class UsesFramework {
-    public List<Integer> ids;
-    public String tag;
+public class PackageInfo {
+    public String forcedPackageId;
+    public String renameManifestPackage;
 }
