+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/apk/VersionInfo.java
@@ -14,7 +14,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package brut.androlib.meta;
+package brut.androlib.apk;
 
 public class VersionInfo {
     public String versionCode;
