+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/ResourcesDecoder.java
@@ -1,0 +1,39 @@
+/*
+ *  Copyright (C) 2010 Ryszard Wiśniewski <brut.alll@gmail.com>
+ *  Copyright (C) 2010 Connor Tumbleson <connor.tumbleson@gmail.com>
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package brut.androlib.res;
+
+import brut.androlib.exceptions.AndrolibException;
+import brut.androlib.Config;
+import brut.directory.*;
+import org.apache.commons.io.IOUtils;
+
+import java.io.*;
+import java.util.*;
+import java.util.logging.Logger;
+
+final public class ResourcesDecoder {
+
+    private final static Logger LOGGER = Logger.getLogger(ResourcesDecoder.class.getName());
+
+    private final Config mConfig;
+
+    public ResourcesDecoder(Config config) {
+        mConfig = config;
+    }
+
+
+}
