+++ brut.apktool/apktool-lib/src/test/java/brut/androlib/aapt1/ReferenceVersionCodeTest.java
@@ -16,11 +16,10 @@
  */
 package brut.androlib.aapt1;
 
-import brut.androlib.ApkBuilder;
 import brut.androlib.ApkDecoder;
 import brut.androlib.BaseTest;
 import brut.androlib.TestUtils;
-import brut.androlib.meta.MetaInfo;
+import brut.androlib.apk.ApkInfo;
 import brut.directory.ExtFile;
 import brut.common.BrutException;
 import brut.util.OS;
@@ -57,7 +56,7 @@ public class ReferenceVersionCodeTest extends BaseTest {
         File outDir = new File(sTmpDir + File.separator + apk + ".out");
         apkDecoder.decode(outDir);
 
-        MetaInfo metaInfo = MetaInfo.readMetaFile(decodedApk);
-        assertEquals("v1.0.0", metaInfo.versionInfo.versionName);
+        ApkInfo apkInfo = ApkInfo.load(decodedApk);
+        assertEquals("v1.0.0", apkInfo.versionInfo.versionName);
     }
 }
